better j2me compilability (44 errors now)
git-svn-id: https://only.mawhrin.net/repos/FBReaderJ/trunk@596 6a642e6f-84f6-412e-ac94-c4a38d5a04b0

+++ src/org/fbreader/fbreader/CollectionView.java
@@ -203,7 +203,7 @@ public class CollectionView extends FBView {
 		
 		public int paragraphNumberByBook(BookDescription book) {
 			Integer num = (Integer)myBookToParagraph.get(book);
-			return (num != null) ? num : 0;
+			return (num != null) ? num.intValue() : 0;
 		}
 
 		public void update() {
